Moved files out of assets folder and renamed main_menu.py to main.py
